+++ routes/api/internal/report.js
@@ -1,0 +1,22 @@
+'use strict';
+
+/**
+ * Endpoint to log csp-reports
+ */
+
+module.exports = function (app) {
+    var logger = app.get('logger');
+
+    app.post('/api/internal/report', function (req, res) {
+
+        logger.error(
+            'CSP report',
+            {
+                headers: req.headers, 
+                body: req.body            
+            }
+        );
+
+        res.ok();
+    });
+};